I made some big progess tonight.  Added the functionality when "buying" a pixel, to be able to check if that pixel is in the database or not, and if not, add it to the requested pixels.
I'm currently stuck behind trying to upload the image to the Firebase storage.

I'll also need to check better for each pixel being bought, instead of just 1.
Also I'll need to work better with the mail api.  Hoping to upload to the storage, and then grab the url from that.  Then I can approve of it or not by email.
Then if approved, I can grab the x,y from it, and add it in, and stuff
Otherwise delete the request image and x,y