Huge update!  Feel like I'm almost on par with where I was at the start of this last Thursday in class.  BUT WITH FIREBASE!
I can't send an email just yet, so that slows me down.  But what I can do is tomorrrow,
I can use robtallguy to write a small php that sends emails.  That should help

My program still takes forever to load, but I put that on my end.  Basically the flow is,
create account
request a square
I approve of said square
database checks if the square is empty, if not, then adds it.

I want to eventually put the paypal option in.  It would fit between checking if the square is empty, reversing the square, then allowing the customer to pay, then adding the square.  I could put in a check to see if any of the squares are unpaid after a day, and then delete them if they aren't.  Or something.